Create insanitized.raceeffect
A .raceeffect file for the NPC-only race in my mod, the Sanitized. 
This gives them lore-fitting strengths and weaknesses, uniquely suited for their nature as "undead", and makes their melee weapons somewhat more dangerous.

While enemy Sanitized are only found on a few planets, it is possible to get Sanitized tenants, so they have biome-based stats even if they don't naturally appear on those planets.

Because they are not playable and likely never will be, I won't be making them an associated .species.patch file.